+++ erasmus/cogs/bible/cog.py
@@ -32,7 +32,7 @@ from ...ui_pages import UIPages
 from ...utils import send_passage
 from .admin_group import BibleAdminGroup
 from .bible_lookup import _BibleOption, bible_lookup
-from .daily_bread_group import DailyBreadGroup, TestingServerPreferencesGroup
+from .daily_bread_group import DailyBreadGroup
 from .search_page_source import SearchPageSource
 from .server_preferences_group import ServerPreferencesGroup
 from .version_group import VersionGroup
@@ -76,7 +76,6 @@ class Bible(Cog['Erasmus']):
     version = VersionGroup()
     server_preferences = ServerPreferencesGroup()
     daily_bread = DailyBreadGroup()
-    testing = TestingServerPreferencesGroup()
 
     __lookup_cooldown: commands.CooldownMapping[discord.Message]
     __daily_bread_task: tasks.Loop[Callable[[], Coroutine[None]]]
@@ -93,7 +92,6 @@ class Bible(Cog['Erasmus']):
         self.admin.initialize_from_parent(self)
         self.version.initialize_from_parent(self)
         self.server_preferences.initialize_from_parent(self)
-        self.testing.initialize_from_parent(self)
         self.daily_bread.initialize_from_parent(self)
 
     async def refresh(self, session: AsyncSession, /) -> None: